Folded BasicLineItemProcessor into CostAndUsageReportLineItemProcessor
