+++ src/java/com/netflix/ice/processor/kubernetes/KubernetesReport.java
@@ -38,8 +38,8 @@ import com.amazonaws.services.s3.model.S3ObjectSummary;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.netflix.ice.common.AwsUtils;
-import com.netflix.ice.common.LineItem;
 import com.netflix.ice.common.ResourceService;
+import com.netflix.ice.processor.LineItem;
 import com.netflix.ice.processor.Report;
 import com.netflix.ice.processor.config.S3BucketConfig;
 import com.netflix.ice.processor.config.KubernetesConfig;
