+++ src/test/com/netflix/ice/basic/BasicResourceServiceTest.java
@@ -36,12 +36,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.netflix.ice.common.LineItem;
 import com.netflix.ice.common.ProductService;
 import com.netflix.ice.common.ResourceService;
 import com.netflix.ice.common.TagConfig;
 import com.netflix.ice.processor.CostAndUsageReport;
 import com.netflix.ice.processor.CostAndUsageReportLineItem;
+import com.netflix.ice.processor.LineItem;
 import com.netflix.ice.processor.config.BillingDataConfig;
 import com.netflix.ice.tag.Account;
 import com.netflix.ice.tag.Product;
